+++ tests/mat2.rs
@@ -2,7 +2,7 @@
 mod support;
 
 macro_rules! impl_mat2_tests {
-    ($t:ident, $const_new:ident, $newmat2:ident, $mat2:ident, $mat3:ident, $newvec2:ident, $vec2:ident) => {
+    ($t:ident, $newmat2:ident, $mat2:ident, $mat3:ident, $newvec2:ident, $vec2:ident) => {
         const IDENTITY: [[$t; 2]; 2] = [[1.0, 0.0], [0.0, 1.0]];
 
         const MATRIX: [[$t; 2]; 2] = [[1.0, 2.0], [3.0, 4.0]];
@@ -17,18 +17,6 @@ macro_rules! impl_mat2_tests {
             assert_eq!(MATRIX1D, M0.to_cols_array());
             assert_eq!(MATRIX1D, M1.to_cols_array());
             assert_eq!(MATRIX1D, M2.to_cols_array());
-        });
-
-        glam_test!(test_const_macros, {
-            #[allow(deprecated)]
-            const M0: $mat2 = $const_new!([0.0; 4]);
-            #[allow(deprecated)]
-            const M1: $mat2 = $const_new!([1.0, 2.0, 3.0, 4.0]);
-            #[allow(deprecated)]
-            const M2: $mat2 = $const_new!([1.0, 2.0], [3.0, 4.0]);
-            assert_eq!($mat2::ZERO, M0);
-            assert_eq!($mat2::from_cols_array(&[1.0, 2.0, 3.0, 4.0]), M1);
-            assert_eq!($mat2::from_cols_array(&[1.0, 2.0, 3.0, 4.0]), M2);
         });
 
         glam_test!(test_mat2_identity, {
@@ -252,8 +240,7 @@ macro_rules! impl_as_ref_tests {
 
 mod mat2 {
     use super::support::deg;
-    #[allow(deprecated)]
-    use glam::{const_mat2, mat2, swizzles::*, vec2, Mat2, Mat3, Vec2};
+    use glam::{mat2, swizzles::*, vec2, Mat2, Mat3, Vec2};
 
     glam_test!(test_align, {
         use std::mem;
@@ -284,14 +271,13 @@ mod mat2 {
         );
     });
 
-    impl_mat2_tests!(f32, const_mat2, mat2, Mat2, Mat3, vec2, Vec2);
+    impl_mat2_tests!(f32, mat2, Mat2, Mat3, vec2, Vec2);
     impl_as_ref_tests!(Mat2);
 }
 
 mod dmat2 {
     use super::support::deg;
-    #[allow(deprecated)]
-    use glam::{const_dmat2, dmat2, dvec2, swizzles::*, DMat2, DMat3, DVec2};
+    use glam::{dmat2, dvec2, swizzles::*, DMat2, DMat3, DVec2};
 
     glam_test!(test_align, {
         use std::mem;
@@ -299,6 +285,6 @@ mod dmat2 {
         assert_eq!(mem::align_of::<DVec2>(), mem::align_of::<DMat2>());
     });
 
-    impl_mat2_tests!(f64, const_dmat2, dmat2, DMat2, DMat3, dvec2, DVec2);
+    impl_mat2_tests!(f64, dmat2, DMat2, DMat3, dvec2, DVec2);
     impl_as_ref_tests!(DMat2);
 }